refactor(client/architecture): modularize TaskBoard components with feature-based structure
