--- client/src/components/TaskForm/index.jsx
@@ -1,2 +1,0 @@
-import React from "react";
-export { default } from "./TaskForm";
