+++ client/src/features/TaskBoard/components/TaskBoardContainer.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import TaskBoard from "../index";
+
+/**
+ * This is the main container for the TaskBoard feature.
+ * It serves as the entry point for the TaskBoard functionality.
+ */
+export function TaskBoardContainer() {
+  return <TaskBoard />;
+}
+
+export default TaskBoardContainer;
