+++ client/src/features/TaskBoard/constants.js
@@ -1,0 +1,4 @@
+// Define item types for drag and drop
+export const ItemTypes = {
+  TASK: "task",
+};
